requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
binance-connector==1.11.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
constantly==15.1.0
cryptography==36.0.1
Django==4.0.3
djangorestframework==3.13.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2021.3
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.8
zope.interface==5.4.0
